test/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
Django==5.1.3
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.2
typing_extensions==4.12.2

